Move utilities out of main code and update XBT_line information retrieval to only source the AODN vocabularies and extract the line description from here.
