+++ SOOP/SOOP_XBT/DELAYED/xbt_config
@@ -348,170 +348,6 @@ SC=Scripps Institution of Oceanography,
 AO=Australian Ocean Data Network,
 NO=NODC,
 
-# lines description
-[IX01]
-title               =Upper Ocean Thermal Data collected on the line IX01 (Fremantle-Sunda Strait) using XBT (expendable bathythermographs)
-XBT_line_description=Fremantle-Sunda-Strait
-XBT_line            =IX01
-
-[indian-ocean/Indian-Ocean]
-title               =Upper Ocean Thermal Data collected in the Indian Ocean using XBT (expendable bathythermographs)
-XBT_line            =Indian-Ocean
-XBT_line_description=Indian-Ocean
-
-[IX12]
-title               =Upper Ocean Thermal Data collected on the line IX12 (Fremantle-Red Sea) using XBT (expendable bathythermographs)
-XBT_line            =IX12
-XBT_line_description=Fremantle-Red-Sea
-
-[IX22/PX11]
-title               =Upper Ocean Thermal Data collected on the line IX22-PX11 (Port Hedland-Japan) using XBT (expendable bathythermographs)
-XBT_line_description=Port-Hedland-Japan
-XBT_line            =IX22-PX11
-
-[PX11/IX22]
-title               =Upper Ocean Thermal Data collected on the line IX22-PX11 (Port Hedland-Japan) using XBT (expendable bathythermographs)
-XBT_line_description=Port-Hedland-Japan
-XBT_line            =IX22-PX11
-
-[PX02]
-title               =Upper Ocean Thermal Data collected on the line PX2 (Across the Banda Sea) using XBT (expendable bathythermographs)
-XBT_line_description=Flores-Sea-Torres-Strait
-XBT_line            =PX02
-
-[PX30/PX31]
-title               =Upper Ocean Thermal Data collected on the high density line PX30-31 (Brisbane-Noumea-Suva/Lautoka) using XBT (expendable bathythermographs)
-XBT_line_description=Brisbane-Noumea-Suva
-XBT_line            =PX30-31
-
-[PX32]
-title               =Upper Ocean Thermal Data collected on the line PX32 (Sydney - Auckland) using XBT (expendable bathythermographs)
-XBT_line_description=Sydney-Auckland
-XBT_line            =PX32
-
-[PX34]
-title               =Upper Ocean Thermal Data collected on the high density line PX34 (Sydney-Wellington) using XBT (expendable bathythermographs)
-XBT_line_description=Sydney-Wellington
-XBT_line            =PX34
-
-[IX09]
-title               =Upper Ocean Thermal Data collected on the line IX9 (Fremantle-Persian Gulf) using XBT (expendable bathythermographs)
-XBT_line_description=Fremantle-Persian-Gulf
-XBT_line            =IX09
-
-[Tasman-sea]
-title               =Upper Ocean Thermal Data collected in the Tasman Sea using XBT (expendable bathythermographs)
-XBT_line_description=Tasman-Sea
-XBT_line            =Tasman-Sea
-
-[PM]
-title               =Upper Ocean Thermal Data collected in Port Moresby region using XBT (expendable bathythermographs)
-XBT_line_description=Port Moresby
-
-[NOLINE]
-title               =Upper Ocean Thermal Data collected using XBT (expendable bathythermographs)
-XBT_line_description=No XBT line
-
-[PX03]
-title               =Upper Ocean Thermal Data collected on the line PX3 (Coral Sea) using XBT (expendable bathythermographs)
-XBT_line_description=Coral-Sea
-XBT_line            =PX03
-
-[IX28]
-title               =Upper Ocean Thermal Data collected on the high density line IX28 (Dumont d Urville-Hobart) using XBT (expendable bathythermographs)
-XBT_line_description=Dumont-d-Urville-Hobart
-XBT_line            =IX28
-
-[Southern-Ocean/SO]
-title               =Upper Ocean Thermal Data collected in the Southern Ocean using XBT (expendable bathythermographs)
-XBT_line            =SO
-XBT_line_description=Southern-Ocean
-
-[PX33]
-title               =Upper Ocean Thermal Data collected on the line PX33 (Hobart-Macquarie Island) using XBT (expendable bathythermographs)
-XBT_line_description=Hobart-Macquarie-Island
-XBT_line            =PX33
-
-[IX02]
-title               =Upper Ocean Thermal Data collected on the line IX02 (Cape of Good Hope - Fremantle) using XBT (expendable bathythermographs)
-XBT_line_description=Cape-of-Good-Hope-Fremantle
-XBT_line            =IX02
-
-[IX21]
-title               =Upper Ocean Thermal Data collected on the line IX21 (Cape of Good Hope - Mauritius) using XBT (expendable bathythermographs)
-XBT_line_description=Cape-of-Good-Hope-Mauritius
-XBT_line            =IX21
-
-[IX15]
-title               =Upper Ocean Thermal Data collected on the line IX15 (Mauritius - Fremantle) using XBT (expendable bathythermographs)
-XBT_line_description=Mauritius-Fremantle
-XBT_line            =IX15
-
-[IX15/IX31]
-title               =Upper Ocean Thermal Data collected on the line IX15-IX31 (Mauritius - Melbourne) using XBT (expendable bathythermographs)
-XBT_line_description=Mauritius-Melbourne
-XBT_line            =IX15-IX31
-
-[IX21/IX06]
-title               =Upper Ocean Thermal Data collected on the line IX21-IX06 (Cape of Good Hope - Mauritius - Malacca Strait) using XBT (expendable bathythermographs)
-XBT_line_description=Cape-of-Good-Hope-Mauritius-Malacca-Strait
-XBT_line            =IX21-IX06
-
-[PX13]
-title               =Upper Ocean Thermal Data collected on the line PX13 (New Zealand - California) using XBT (expendable bathythermographs)
-XBT_line_description=New-Zealand-California
-XBT_line            =PX13
-
-[PX31]
-title               =Upper Ocean Thermal Data collected on the line PX31 (Noumea - Suva) using XBT (expendable bathythermographs)
-XBT_line_description=Noumea-Suva
-XBT_line            =PX31
-
-[PX06]
-title               =Upper Ocean Thermal Data collected on the line PX06 (Suva - Auckland) using XBT (expendable bathythermographs)
-XBT_line_description=Suva-Auckland
-XBT_line            =PX06
-
-[IX22]
-title               =Upper Ocean Thermal Data collected on the line IX22 (Fremantle-Flores Sea) using XBT (expendable bathythermographs)
-XBT_line_description=Fremantle-Flores-Sea
-XBT_line            =IX22
-
-[PX05]
-title               =Upper Ocean Thermal Data collected on the line PX5 (Brisbane-Japan) using XBT (expendable bathythermographs)
-XBT_line_description=Brisbane-Japan and Japan-New Zealand
-XBT_line            =PX05
-
-[IX08]
-title               =Upper Ocean Thermal Data collected on the line IX8 (Mauritius - Bombay) using XBT (expendable bathythermographs)
-XBT_line_description=Mauritius-Bombay
-XBT_line            =IX08
-
-[PX57]
-title               =Upper Ocean Thermal Data collected on the line PX57 (Brisbane-Wellington) using XBT (expendable bathythermographs)
-XBT_line_description=Brisbane-Wellington
-XBT_line            =PX57
-
-[PX28]
-title               =Upper Ocean Thermal Data collected on the line PX28 (Tahiti-Auckland) using XBT (expendable bathythermographs)
-XBT_line_description=Tahiti-Auckland
-XBT_line            =PX28
-
-[PX17]
-title               =Upper Ocean Thermal Data collected on the line PX17 (Tahiti-Panama) using XBT (expendable bathythermographs)
-XBT_line_description=Tahiti-Panama
-XBT_line            =PX17
-
-[PX35]
-title               =Upper Ocean Thermal Data collected on the line PX35 (Melbourne-Dunedin) using XBT (expendable bathythermographs)
-XBT_line_description=Melbourne-Dunedin
-XBT_line            =PX35
-
-[PX55]
-title               =Upper Ocean Thermal Data collected on the line PX55 (Melbourne-Wellington) using XBT (expendable bathythermographs)
-XBT_line_description=Melbourne-Wellington
-XBT_line            =PX55
-
 # Filename prefix to allow for different datasets, not just IMOS
 # IMOS_SOOP-
 # Cruise_ID (for other sources (eg, Investigator, Southern Surveyor))
